:art: move public static util methods from `InstagramPostsAdapter` to `Utils.java`
